Cleaned up analytics module structure
