+++ public/main.js
@@ -1,8 +1,3 @@
-import analytics from './components/analytics/analytics';
+import { logPageView } from './components/analytics/analytics';
 
-import { init as initOphan } from './components/analytics/ophan';
-
-analytics.go();
-
-initOphan();
-
+logPageView();
